Manually set up a Shiny app to use overshiny

overshiny will set up automatically if you have an overlayPlotOutput() anywhere in your Shiny UI, which you probably do if you are using this package. But if you don't, you can set up overshiny by manually putting useOverlay() somewhere in your Shiny app's UI.

Usage

useOverlay()

Value

Returns an HTML dependency that sets up your Shiny app to use overshiny.

Details

This also calls shinyjs::useShinyjs(), as overshiny depends on shinyjs.

See also

Examples

ui <- shiny::fluidPage(
    useOverlay() # only needed if no overlayPlotOutput() elements below
    # further UI elements here . . .
)

server <- function(input, output) {
    # server code here . . .
}

if (interactive()) {
    shiny::shinyApp(ui, server)
}